add places route
